- Switching to float precision. - Minor version bump. - Updated test cases.

+++ src/main/java/io/vacco/jtinn/activation/JtActivationFn.java
@@ -3,6 +3,6 @@ package io.vacco.jtinn.activation;
 import java.io.Serializable;
 
 public interface JtActivationFn extends Serializable {
-  double apply(double z);
-  double pd(double z);
+  float apply(float z);
+  float pd(float z);
 }
