+++ src/main/java/io/vacco/jtinn/activation/JtSigmoid.java
@@ -6,11 +6,11 @@ public class JtSigmoid implements JtActivationFn {
 
   private static final long serialVersionUID = JtSchema.version;
 
-  @Override public double apply(double z) {
-    return 1.0 / (1.0 + Math.exp(-z));
+  @Override public float apply(float z) {
+    return (float) (1.0 / (1.0 + Math.exp(-z)));
   }
 
-  @Override public double pd(double z) {
-    return z * (1.0 - z);
+  @Override public float pd(float z) {
+    return (float) (z * (1.0 - z));
   }
 }
