+++ src/main/java/io/vacco/jtinn/net/JtPredictionSample.java
@@ -2,10 +2,10 @@ package io.vacco.jtinn.net;
 
 public class JtPredictionSample {
 
-  public double[] features;
-  public double[] labels;
+  public float[] features;
+  public float[] labels;
 
-  public static JtPredictionSample of(double[] features, double[] labels) {
+  public static JtPredictionSample of(float[] features, float[] labels) {
     JtPredictionSample sample = new JtPredictionSample();
     sample.features = features;
     sample.labels = labels;
