+++ src/main/java/io/vacco/jtinn/net/JtSgdUpdater.java
@@ -4,19 +4,19 @@ public class JtSgdUpdater implements JtUpdater {
 
   private static final long serialVersionUID = JtSchema.version;
 
-  public double learningRate, annealingRate;
+  public float learningRate, annealingRate;
 
-  public JtSgdUpdater init(double learningRate, double annealingRate) {
+  public JtSgdUpdater init(float learningRate, float annealingRate) {
     this.learningRate = learningRate;
     this.annealingRate = annealingRate;
     return this;
   }
 
-  @Override public void apply(double[] lm1a, JtLayer l) {
+  @Override public void apply(float[] lm1a, JtLayer l) {
     for (int j = 0; j < l.size(); j++) {
       for (int w = 0; w < l.weightSize(); w++) {
         double pda = l instanceof JtOutputLayer ? lm1a[w] : l.actFn.pd(l.a[j]) * lm1a[w];
-        l.w[j][w] = l.w[j][w] - (learningRate * l.δ[j] * pda);
+        l.w[j][w] = (float) (l.w[j][w] - (learningRate * l.δ[j] * pda));
       }
       l.b[j] = l.b[j] - l.δ[j];
     }
